Backend/requirements.txt
flask
flask-cors
web3
opencv-python
pillow
torch
facenet-pytorch
numpy
